(tao) % Practice non-action. Work without doing. Taste the tasteless. Magnify the small, increase the few. Reward bitterness with care.
See simplicity in the complicated.
Achieve greatness in little things.

In the universe the difficult things are done as if they are easy.
In the universe great acts are made up of small deeds.
The sage does not attempt anything very big,
And thus achieved greatness.

Easy promises make for little trust.
Taking things lightly results in great difficulty.
Because the sage always confronts difficulties,
He never experiences them.